organize
https://github.com/metasfresh/metasfresh-webui-api/issues/1148
