+++ src/main/java/de/metas/ui/web/order/products_proposal/filters/ProductsProposalViewFilters.java
@@ -1,4 +1,4 @@
-package de.metas.ui.web.order.products_proposal.view;
+package de.metas.ui.web.order.products_proposal.filters;
 
 import java.util.List;
