+++ src/main/java/de/metas/ui/web/order/products_proposal/process/WEBUI_ProductsProposal_AddProductFromBasePriceList.java
@@ -3,7 +3,7 @@ package de.metas.ui.web.order.products_proposal.process;
 import java.util.List;
 
 import de.metas.process.ProcessPreconditionsResolution;
-import de.metas.ui.web.order.products_proposal.view.ProductsProposalRow;
+import de.metas.ui.web.order.products_proposal.model.ProductsProposalRow;
 import de.metas.ui.web.order.products_proposal.view.ProductsProposalView;
 
 /*
