+++ src/main/java/de/metas/ui/web/order/products_proposal/view/OtherSalePricesProductsProposalViewFactory.java
@@ -22,6 +22,9 @@ import de.metas.lang.SOTrx;
 import de.metas.money.MoneyService;
 import de.metas.process.RelatedProcessDescriptor;
 import de.metas.product.ProductId;
+import de.metas.ui.web.order.products_proposal.model.ProductsProposalRow;
+import de.metas.ui.web.order.products_proposal.model.ProductsProposalRowsData;
+import de.metas.ui.web.order.products_proposal.model.ProductsProposalRowsLoader;
 import de.metas.ui.web.order.products_proposal.process.WEBUI_ProductsProposal_AddProductFromBasePriceList;
 import de.metas.ui.web.order.products_proposal.process.WEBUI_ProductsProposal_CancelAddingProductFromBasePriceList;
 import de.metas.ui.web.order.products_proposal.process.WEBUI_ProductsProposal_ShowProductsSoldToOtherCustomers;
