+++ src/main/java/de/metas/ui/web/order/products_proposal/view/ProductsProposalView.java
@@ -19,6 +19,11 @@ import de.metas.process.RelatedProcessDescriptor;
 import de.metas.product.ProductId;
 import de.metas.ui.web.document.filter.DocumentFilter;
 import de.metas.ui.web.document.filter.NullDocumentFilterDescriptorsProvider;
+import de.metas.ui.web.order.products_proposal.filters.ProductsProposalViewFilter;
+import de.metas.ui.web.order.products_proposal.filters.ProductsProposalViewFilters;
+import de.metas.ui.web.order.products_proposal.model.ProductsProposalRow;
+import de.metas.ui.web.order.products_proposal.model.ProductsProposalRowChangeRequest;
+import de.metas.ui.web.order.products_proposal.model.ProductsProposalRowsData;
 import de.metas.ui.web.view.AbstractCustomView;
 import de.metas.ui.web.view.IEditableView;
 import de.metas.ui.web.view.ViewId;
